studybud/requirements.txt
django==5.0.3
asgiref
autopep8
django-cors-headers
djangorestframework
Pillow
pycodestyle
pytz
sqlparse
toml
python-dotenv==1.0.1
gunicorn==21.2.0
django-otp==1.3.0
qrcode==7.4.2
whitenoise==6.6.0
channels==4.1.0
daphne==4.1.2
uvicorn==0.29.0